% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_complete_yaml.R
\name{is_complete_yaml}
\alias{is_complete_yaml}
\title{Check if the YAML is complete}
\usage{
is_complete_yaml(
  path,
  n = 50,
  keys = c("title", "author", "date"),
  values = c("___", "", " ", "---", "...", "sv")
)
}
\arguments{
\item{path}{The path of rmd file}

\item{n}{The default number of lines read from the R markdown document}

\item{keys}{The keys to be assessed}

\item{values}{Values considered invalid}
}
\value{
a logical vector
}
\description{
The YAML metadata begins the R markdown document.
This information are use by Pandoc, rmarkdown, and knitr.
It includes several instructions in the form key-value.
The information provided can influence the code, content, and the rendering process.
}
\examples{

#is_complete_yaml(path= "rmd_file")

}
