% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{strb_extract}
\alias{strb_extract}
\title{Extract patterns from a string.}
\usage{
strb_extract(string, pattern)
}
\arguments{
\item{string}{a character vector}

\item{pattern}{Pattern to look for}
}
\value{
A character vector.
}
\description{
strb_extract() is the combination of \code{regexpr()} and \code{regmatches()} with argument to \code{stringr::str_subset()}.
This function is slower and requires more memory than the {stringr} functions.
The {stringr} package depends on more than 50 packages while strb_* uses only R base.
For analyses that do not require a large amount of memory,
the strb_* functions seem to combine the advantages of basic R with the simple syntax of stringr.
}
\seealso{
\link[stringr:str_extract]{stringr::str_extract}, \code{\link[=regexpr]{regexpr()}} and \code{\link[=regmatches]{regmatches()}}
}
