% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.R
\name{strb_replace}
\alias{strb_replace}
\alias{strb_replace_all}
\title{Keep strings matching a pattern.}
\usage{
strb_replace(string, pattern, replacement)

strb_replace_all(string, pattern, replacement)
}
\arguments{
\item{string}{a character vector}

\item{pattern}{Pattern to look for}

\item{replacement}{A character vector of replacements.}
}
\value{
A character vector.
}
\description{
\code{strb_replace()} is \code{sub()} with argument to \code{stringr::str_replace()}.
\code{strb_replace_all()} is \code{gsub()} with argument to \code{stringr::str_replace_all()}.
The {stringr} package depends on more than 50 packages while strb_* uses only R base.
For analyses that do not require a large amount of memory,
the strb_* functions seem to combine the advantages of basic R with the simple syntax of stringr.
}
\examples{
fruit <- c("apple", "banana", "pear", "pineapple")
strb_replace(fruit, "^a", "_")
strb_replace_all(fruit, "^a", "_")

}
\seealso{
\code{\link[stringr:str_replace]{stringr::str_replace()}}, \code{\link[stringr:str_replace]{stringr::str_replace_all()}} \code{\link[=sub]{sub()}} and \code{\link[=gsub]{gsub()}}
}
